Adding json tests
