+++ tests/functional/kafka_consumer/required_dependencies/test_kafka_consumer.py
@@ -1,56 +1,32 @@
 import base64
-import io
 import json
+from copy import deepcopy
 from dataclasses import dataclass
 
 import pytest
-from avro.io import BinaryEncoder, DatumWriter
-from avro.schema import parse as parse_schema
-from pydantic import BaseModel
 
 from aws_lambda_powertools.utilities.kafka_consumer.consumer_records import ConsumerRecords
+from aws_lambda_powertools.utilities.kafka_consumer.exceptions import (
+    KafkaConsumerDeserializationError,
+)
 from aws_lambda_powertools.utilities.kafka_consumer.kafka_consumer import kafka_consumer
 from aws_lambda_powertools.utilities.kafka_consumer.schema_config import SchemaConfig
 
 
 @pytest.fixture
-def avro_schema():
-    return """
-    {
-        "type": "record",
-        "name": "User",
-        "namespace": "com.example",
-        "fields": [
-            {"name": "name", "type": "string"},
-            {"name": "age", "type": "int"}
-        ]
-    }
-    """
+def json_encoded_value():
+    data = {"name": "John Doe", "age": 30}
+    return base64.b64encode(json.dumps(data).encode("utf-8")).decode("utf-8")
 
 
 @pytest.fixture
-def avro_encoded_data():
-    schema_str = """
-    {
-        "type": "record",
-        "name": "User",
-        "namespace": "com.example",
-        "fields": [
-            {"name": "name", "type": "string"},
-            {"name": "age", "type": "int"}
-        ]
-    }
-    """
-    parsed_schema = parse_schema(schema_str)
-    writer = DatumWriter(parsed_schema)
-    bytes_writer = io.BytesIO()
-    encoder = BinaryEncoder(bytes_writer)
-    writer.write({"name": "John Doe", "age": 30}, encoder)
-    return base64.b64encode(bytes_writer.getvalue()).decode("utf-8")
+def json_encoded_key():
+    data = {"user_id": "123"}
+    return base64.b64encode(json.dumps(data).encode("utf-8")).decode("utf-8")
 
 
 @pytest.fixture
-def kafka_event_with_avro_data(avro_encoded_data):
+def kafka_event_with_json_data(json_encoded_value, json_encoded_key):
     return {
         "eventSource": "aws:kafka",
         "eventSourceArn": "arn:aws:kafka:us-east-1:123456789012:cluster/my-cluster/abcdefg",
@@ -62,8 +38,8 @@ def kafka_event_with_avro_data(avro_encoded_data):
                     "offset": 15,
                     "timestamp": 1545084650987,
                     "timestampType": "CREATE_TIME",
-                    "key": None,
-                    "value": avro_encoded_data,
+                    "key": json_encoded_key,
+                    "value": json_encoded_value,
                     "headers": [{"headerKey": [104, 101, 97, 100, 101, 114, 86, 97, 108, 117, 101]}],
                 },
             ],
@@ -71,192 +47,18 @@ def kafka_event_with_avro_data(avro_encoded_data):
     }
 
 
-# Test Models
-
-
-class UserSchema(BaseModel):
+@dataclass
+class UserValueDataClass:
     name: str
     age: int
 
 
-def test_kafka_consumer_with_avro_and_pydantic(kafka_event_with_avro_data, avro_schema, lambda_context):
-    """Test Kafka consumer with Avro deserialization and Pydantic output serialization."""
-
-    # Create temp file to capture results
-    result_data = {}
-
-    schema_config = SchemaConfig(
-        value_schema_type="AVRO",
-        value_schema=avro_schema,
-        value_output_serializer=UserSchema,
-    )
-
-    @kafka_consumer(schema_config=schema_config)
-    def handler(event: ConsumerRecords, context):
-        # Capture the results to verify
-        record = next(event.records)
-        result_data["value_type"] = type(record.value).__name__
-        result_data["name"] = record.value.name
-        result_data["age"] = record.value.age
-        return {"processed": True}
-
-    # Call the handler
-    result = handler(kafka_event_with_avro_data, lambda_context)
-
-    # Verify the results
-    assert result == {"processed": True}
-    assert result_data["value_type"] == "UserSchema"
-    assert result_data["name"] == "John Doe"
-    assert result_data["age"] == 30
+@dataclass
+class UserKeyClass:
+    user_id: str
 
 
-def test_kafka_consumer_with_avro_and_dataclass(kafka_event_with_avro_data, avro_schema, lambda_context):
-    """Test Kafka consumer with Avro deserialization and dataclass output serialization."""
-
-    # Create dict to capture results
-    result_data = {}
-
-    schema_config = SchemaConfig(
-        value_schema_type="AVRO",
-        value_schema=avro_schema,
-        value_output_serializer=UserDataClass,
-    )
-
-    @kafka_consumer(schema_config=schema_config)
-    def handler(event: ConsumerRecords, context):
-        # Capture the results to verify
-        record = next(event.records)
-        result_data["value_type"] = type(record.value).__name__
-        result_data["name"] = record.value.name
-        result_data["age"] = record.value.age
-        return {"processed": True}
-
-    # Call the handler
-    result = handler(kafka_event_with_avro_data, lambda_context)
-
-    # Verify the results
-    assert result == {"processed": True}
-    assert result_data["value_type"] == "UserDataClass"
-    assert result_data["name"] == "John Doe"
-    assert result_data["age"] == 30
-
-
-def test_kafka_consumer_with_avro_and_custom_object(kafka_event_with_avro_data, avro_schema, lambda_context):
-    """Test Kafka consumer with Avro deserialization and custom object serialization."""
-
-    # Create dict to capture results
-    result_data = {}
-
-    schema_config = SchemaConfig(
-        value_schema_type="AVRO",
-        value_schema=avro_schema,
-        value_output_serializer=UserDict,
-    )
-
-    @kafka_consumer(schema_config=schema_config)
-    def handler(event: ConsumerRecords, context):
-        # Capture the results to verify
-        record = next(event.records)
-        result_data["value_type"] = type(record.value).__name__
-        result_data["name"] = record.value.name
-        result_data["age"] = record.value.age
-        return {"processed": True}
-
-    # Call the handler
-    result = handler(kafka_event_with_avro_data, lambda_context)
-
-    # Verify the results
-    assert result == {"processed": True}
-    assert result_data["value_type"] == "UserDict"
-    assert result_data["name"] == "John Doe"
-    assert result_data["age"] == 30
-
-
-def test_kafka_consumer_with_avro_raw(kafka_event_with_avro_data, avro_schema, lambda_context):
-    """Test Kafka consumer with Avro deserialization without output serialization."""
-
-    # Create dict to capture results
-    result_data = {}
-
-    schema_config = SchemaConfig(value_schema_type="AVRO", value_schema=avro_schema)
-
-    @kafka_consumer(schema_config=schema_config)
-    def handler(event: ConsumerRecords, context):
-        # Capture the results to verify
-        record = next(event.records)
-        result_data["value_type"] = type(record.value).__name__
-        result_data["name"] = record.value["name"]
-        result_data["age"] = record.value["age"]
-        return {"processed": True}
-
-    # Call the handler
-    result = handler(kafka_event_with_avro_data, lambda_context)
-
-    # Verify the results
-    assert result == {"processed": True}
-    assert result_data["value_type"] == "dict"
-    assert result_data["name"] == "John Doe"
-    assert result_data["age"] == 30
-
-
-# Tests for Kafka Consumer with JSON Deserializer
-
-
-def test_kafka_consumer_with_json_and_pydantic(kafka_event_with_json_data, lambda_context):
-    """Test Kafka consumer with JSON deserialization and Pydantic output serialization."""
-
-    # Create dict to capture results
-    result_data = {}
-
-    schema_config = SchemaConfig(value_schema_type="JSON", value_output_serializer=UserSchema)
-
-    @kafka_consumer(schema_config=schema_config)
-    def handler(event: ConsumerRecords, context):
-        # Capture the results to verify
-        record = next(event.records)
-        result_data["value_type"] = type(record.value).__name__
-        result_data["name"] = record.value.name
-        result_data["age"] = record.value.age
-        return {"processed": True}
-
-    # Call the handler
-    result = handler(kafka_event_with_json_data, lambda_context)
-
-    # Verify the results
-    assert result == {"processed": True}
-    assert result_data["value_type"] == "UserSchema"
-    assert result_data["name"] == "John Doe"
-    assert result_data["age"] == 30
-
-
-def test_kafka_consumer_with_json_and_dataclass(kafka_event_with_json_data, lambda_context):
-    """Test Kafka consumer with JSON deserialization and dataclass output serialization."""
-
-    # Create dict to capture results
-    result_data = {}
-
-    schema_config = SchemaConfig(value_schema_type="JSON", value_output_serializer=UserDataClass)
-
-    @kafka_consumer(schema_config=schema_config)
-    def handler(event: ConsumerRecords, context):
-        # Capture the results to verify
-        record = next(event.records)
-        result_data["value_type"] = type(record.value).__name__
-        result_data["name"] = record.value.name
-        result_data["age"] = record.value.age
-        return {"processed": True}
-
-    # Call the handler
-    result = handler(kafka_event_with_json_data, lambda_context)
-
-    # Verify the results
-    assert result == {"processed": True}
-    assert result_data["value_type"] == "UserDataClass"
-    assert result_data["name"] == "John Doe"
-    assert result_data["age"] == 30
-
-
-def test_kafka_consumer_with_json_raw(kafka_event_with_json_data, lambda_context):
+def test_kafka_consumer_with_json(kafka_event_with_json_data, lambda_context):
     """Test Kafka consumer with JSON deserialization without output serialization."""
 
     # Create dict to capture results
@@ -283,161 +85,13 @@ def test_kafka_consumer_with_json_raw(kafka_event_with_json_data, lambda_context
     assert result_data["age"] == 30
 
 
-# Tests for Error Handling
-
-
-def test_kafka_consumer_with_avro_missing_schema():
-    """Test error handling when Avro schema is missing."""
-
-    with pytest.raises(ValueError) as excinfo:
-        # This will raise validation error in SchemaConfig init
-        SchemaConfig(value_schema_type="AVRO", value_schema=None)
-
-    assert "value_schema_str must be provided" in str(excinfo.value)
-
-
-def test_kafka_consumer_with_invalid_avro_data(lambda_context):
-    """Test error handling when Avro data is invalid."""
-
-    avro_schema = """
-    {
-        "type": "record",
-        "name": "User",
-        "namespace": "com.example",
-        "fields": [
-            {"name": "name", "type": "string"},
-            {"name": "age", "type": "int"}
-        ]
-    }
-    """
-
-    # Create invalid avro data
-    invalid_data = base64.b64encode(b"invalid avro data").decode("utf-8")
-
-    kafka_event = {
-        "eventSource": "aws:kafka",
-        "records": {
-            "my-topic-1": [
-                {
-                    "topic": "my-topic-1",
-                    "partition": 0,
-                    "offset": 15,
-                    "timestamp": 1545084650987,
-                    "timestampType": "CREATE_TIME",
-                    "key": None,
-                    "value": invalid_data,
-                    "headers": [],
-                },
-            ],
-        },
-    }
-
-    schema_config = SchemaConfig(value_schema_type="AVRO", value_schema=avro_schema)
-
-    @kafka_consumer(schema_config=schema_config)
-    def handler(event: ConsumerRecords, context):
-        # This should never be reached if deserializer fails
-        return {"processed": True}
-
-    # This should raise a deserialization error
-    with pytest.raises(Exception) as excinfo:
-        handler(kafka_event, lambda_context)
-
-    # The exact error message may vary depending on the Avro library's internals,
-    # but should indicate a deserialization problem
-    assert "deseriali" in str(excinfo.value).lower() or "avro" in str(excinfo.value).lower()
-
-
-def test_kafka_consumer_with_invalid_json_data(lambda_context):
-    """Test error handling when JSON data is invalid."""
-
-    # Create invalid JSON data
-    invalid_data = base64.b64encode(b"invalid json data").decode("utf-8")
-
-    kafka_event = {
-        "eventSource": "aws:kafka",
-        "records": {
-            "my-topic-1": [
-                {
-                    "topic": "my-topic-1",
-                    "partition": 0,
-                    "offset": 15,
-                    "timestamp": 1545084650987,
-                    "timestampType": "CREATE_TIME",
-                    "key": None,
-                    "value": invalid_data,
-                    "headers": [],
-                },
-            ],
-        },
-    }
-
-    schema_config = SchemaConfig(value_schema_type="JSON")
-
-    @kafka_consumer(schema_config=schema_config)
-    def handler(event: ConsumerRecords, context):
-        # This should never be reached if deserializer fails
-        return {"processed": True}
-
-    # This should raise a deserialization error
-    with pytest.raises(Exception) as excinfo:
-        handler(kafka_event, lambda_context)
-
-    assert "json" in str(excinfo.value).lower() or "deseriali" in str(excinfo.value).lower()
-
-
-# Tests for Complex Types with Pydantic TypeAdapter
-
-
-def test_kafka_consumer_with_complex_pydantic_models(kafka_event_with_json_data, lambda_context):
-    """Test handling complex nested Pydantic models."""
-
-    # Prepare a complex event with nested structure
-    complex_data = {
-        "name": "John Doe",
-        "age": 30,
-        "address": {"street": "123 Main St", "city": "Anytown", "zip": "12345"},
-        "tags": ["customer", "premium"],
-    }
-
-    # Encode the complex data
-    complex_encoded = base64.b64encode(json.dumps(complex_data).encode("utf-8")).decode("utf-8")
-
-    # Create a kafka event with the complex data
-    complex_event = {
-        "eventSource": "aws:kafka",
-        "records": {
-            "my-topic-1": [
-                {
-                    "topic": "my-topic-1",
-                    "partition": 0,
-                    "offset": 15,
-                    "timestamp": 1545084650987,
-                    "timestampType": "CREATE_TIME",
-                    "key": None,
-                    "value": complex_encoded,
-                    "headers": [],
-                },
-            ],
-        },
-    }
-
-    # Define complex pydantic models
-    class Address(BaseModel):
-        street: str
-        city: str
-        zip: str
-
-    class ComplexUser(BaseModel):
-        name: str
-        age: int
-        address: Address
-        tags: list[str]
+def test_kafka_consumer_with_json_and_dataclass(kafka_event_with_json_data, lambda_context):
+    """Test Kafka consumer with JSON deserialization and dataclass output serialization."""
 
     # Create dict to capture results
     result_data = {}
 
-    schema_config = SchemaConfig(value_schema_type="JSON", value_output_serializer=ComplexUser)
+    schema_config = SchemaConfig(value_schema_type="JSON", value_output_serializer=UserValueDataClass)
 
     @kafka_consumer(schema_config=schema_config)
     def handler(event: ConsumerRecords, context):
@@ -446,148 +100,43 @@ def test_kafka_consumer_with_complex_pydantic_models(kafka_event_with_json_data,
         result_data["value_type"] = type(record.value).__name__
         result_data["name"] = record.value.name
         result_data["age"] = record.value.age
-        result_data["street"] = record.value.address.street
-        result_data["city"] = record.value.address.city
-        result_data["zip"] = record.value.address.zip
-        result_data["tags"] = record.value.tags
         return {"processed": True}
 
     # Call the handler
-    result = handler(complex_event, lambda_context)
+    result = handler(kafka_event_with_json_data, lambda_context)
 
     # Verify the results
     assert result == {"processed": True}
-    assert result_data["value_type"] == "ComplexUser"
+    assert result_data["value_type"] == "UserValueDataClass"
     assert result_data["name"] == "John Doe"
     assert result_data["age"] == 30
-    assert result_data["street"] == "123 Main St"
-    assert result_data["city"] == "Anytown"
-    assert result_data["zip"] == "12345"
-    assert "customer" in result_data["tags"]
-    assert "premium" in result_data["tags"]
 
 
-def test_kafka_consumer_with_union_types(lambda_context):
-    """Test handling Pydantic models with union types using TypeAdapter."""
+def test_kafka_consumer_with_invalid_json_data(kafka_event_with_json_data, lambda_context):
+    """Test error handling when JSON data is invalid."""
 
-    # Define two possible payload types
-    class TextMessage(BaseModel):
-        type: str = "text"
-        content: str
-        timestamp: int
+    # Create invalid JSON data
+    invalid_data = "invalid json data"
+    kafka_event_with_json_data = deepcopy(kafka_event_with_json_data)
+    kafka_event_with_json_data["records"]["my-topic-1"][0]["value"] = invalid_data
 
-    class ImageMessage(BaseModel):
-        type: str = "image"
-        url: str
-        width: int
-        height: int
-        timestamp: int
-
-    # Define a union type
-    MessageType = TextMessage | ImageMessage
-
-    # Create a text message event
-    text_data = {"type": "text", "content": "Hello world", "timestamp": 1636718400000}
-
-    text_encoded = base64.b64encode(json.dumps(text_data).encode("utf-8")).decode("utf-8")
-
-    text_event = {
-        "eventSource": "aws:kafka",
-        "records": {
-            "my-topic-1": [
-                {
-                    "topic": "my-topic-1",
-                    "partition": 0,
-                    "offset": 15,
-                    "timestamp": 1545084650987,
-                    "timestampType": "CREATE_TIME",
-                    "key": None,
-                    "value": text_encoded,
-                    "headers": [],
-                },
-            ],
-        },
-    }
-
-    # Create an image message event
-    image_data = {
-        "type": "image",
-        "url": "https://example.com/image.jpg",
-        "width": 800,
-        "height": 600,
-        "timestamp": 1636718500000,
-    }
-
-    image_encoded = base64.b64encode(json.dumps(image_data).encode("utf-8")).decode("utf-8")
-
-    image_event = {
-        "eventSource": "aws:kafka",
-        "records": {
-            "my-topic-1": [
-                {
-                    "topic": "my-topic-1",
-                    "partition": 0,
-                    "offset": 16,
-                    "timestamp": 1545084650987,
-                    "timestampType": "CREATE_TIME",
-                    "key": None,
-                    "value": image_encoded,
-                    "headers": [],
-                },
-            ],
-        },
-    }
-
-    # Create dict to capture results for text
-    text_result = {}
-
-    # Create schema config with union type
-    schema_config = SchemaConfig(value_schema_type="JSON", value_output_serializer=MessageType)
+    schema_config = SchemaConfig(value_schema_type="JSON")
 
     @kafka_consumer(schema_config=schema_config)
-    def text_handler(event: ConsumerRecords, context):
+    def handler(event: ConsumerRecords, context):
+        # This should never be reached if deserializer fails
         record = next(event.records)
-        text_result["type"] = type(record.value).__name__
-        text_result["message_type"] = record.value.type
-        text_result["content"] = record.value.content
-        text_result["timestamp"] = record.value.timestamp
+        assert record.value
         return {"processed": True}
 
-    # Call the handler with text message
-    text_handler(text_event, lambda_context)
+    # This should raise a deserialization error
+    with pytest.raises(KafkaConsumerDeserializationError) as excinfo:
+        handler(kafka_event_with_json_data, lambda_context)
 
-    # Verify text message results
-    assert text_result["type"] == "TextMessage"
-    assert text_result["message_type"] == "text"
-    assert text_result["content"] == "Hello world"
-    assert text_result["timestamp"] == 1636718400000
-
-    # Create dict to capture results for image
-    image_result = {}
-
-    @kafka_consumer(schema_config=schema_config)
-    def image_handler(event: ConsumerRecords, context):
-        record = next(event.records)
-        image_result["type"] = type(record.value).__name__
-        image_result["message_type"] = record.value.type
-        image_result["url"] = record.value.url
-        image_result["width"] = record.value.width
-        image_result["height"] = record.value.height
-        image_result["timestamp"] = record.value.timestamp
-        return {"processed": True}
-
-    # Call the handler with image message
-    image_handler(image_event, lambda_context)
-
-    # Verify image message results
-    assert image_result["type"] == "ImageMessage"
-    assert image_result["message_type"] == "image"
-    assert image_result["url"] == "https://example.com/image.jpg"
-    assert image_result["width"] == 800
-    assert image_result["height"] == 600
-    assert image_result["timestamp"] == 1636718500000
+    assert "Error trying to deserialize json data" in str(excinfo.value)
 
 
+# Tests for Complex Types with Pydantic TypeAdapter
 def test_kafka_consumer_with_multiple_records(lambda_context):
     """Test processing multiple records in a single event."""
 
@@ -641,7 +190,7 @@ def test_kafka_consumer_with_multiple_records(lambda_context):
     }
 
     # Create schema config
-    schema_config = SchemaConfig(value_schema_type="JSON", value_output_serializer=UserSchema)
+    schema_config = SchemaConfig(value_schema_type="JSON", value_output_serializer=UserValueDataClass)
 
     # Create list to store processed records
     processed_records = []
@@ -664,169 +213,6 @@ def test_kafka_consumer_with_multiple_records(lambda_context):
     assert any(r["name"] == "Bob Johnson" and r["age"] == 40 for r in processed_records)
 
 
-def test_kafka_consumer_with_key_deserialization(lambda_context, avro_schema):
-    """Test deserializing both key and value with different schemas and serializers."""
-
-    # Key schema for Avro
-    key_schema = """
-    {
-        "type": "record",
-        "name": "Key",
-        "namespace": "com.example",
-        "fields": [
-            {"name": "id", "type": "string"}
-        ]
-    }
-    """
-
-    # Create Avro encoded key
-    key_data = {"id": "user-123"}
-    parsed_key_schema = parse_schema(key_schema)
-    writer = DatumWriter(parsed_key_schema)
-    bytes_writer = io.BytesIO()
-    encoder = BinaryEncoder(bytes_writer)
-    writer.write(key_data, encoder)
-    key_encoded = base64.b64encode(bytes_writer.getvalue()).decode("utf-8")
-
-    # Create Avro encoded value
-    value_data = {"name": "John Doe", "age": 30}
-    parsed_value_schema = parse_schema(avro_schema)
-    writer = DatumWriter(parsed_value_schema)
-    bytes_writer = io.BytesIO()
-    encoder = BinaryEncoder(bytes_writer)
-    writer.write(value_data, encoder)
-    value_encoded = base64.b64encode(bytes_writer.getvalue()).decode("utf-8")
-
-    # Create a kafka event with key and value
-    event = {
-        "eventSource": "aws:kafka",
-        "records": {
-            "my-topic-1": [
-                {
-                    "topic": "my-topic-1",
-                    "partition": 0,
-                    "offset": 15,
-                    "timestamp": 1545084650987,
-                    "timestampType": "CREATE_TIME",
-                    "key": key_encoded,
-                    "value": value_encoded,
-                    "headers": [],
-                },
-            ],
-        },
-    }
-
-    # Define key model
-    class KeyModel(BaseModel):
-        id: str
-
-    # Create dict to capture results
-    key_value_result = {}
-
-    # Create schema config with both key and value
-    schema_config = SchemaConfig(
-        value_schema_type="AVRO",
-        value_schema=avro_schema,
-        value_output_serializer=UserSchema,
-        key_schema_type="AVRO",
-        key_schema=key_schema,
-        key_output_serializer=KeyModel,
-    )
-
-    @kafka_consumer(schema_config=schema_config)
-    def handler(event: ConsumerRecords, context):
-        record = next(event.records)
-        key_value_result["key_type"] = type(record.key).__name__
-        key_value_result["key_id"] = record.key.id
-        key_value_result["value_type"] = type(record.value).__name__
-        key_value_result["value_name"] = record.value.name
-        key_value_result["value_age"] = record.value.age
-        return {"processed": True}
-
-    # Call the handler
-    result = handler(event, lambda_context)
-
-    # Verify the results
-    assert result == {"processed": True}
-    assert key_value_result["key_type"] == "KeyModel"
-    assert key_value_result["key_id"] == "user-123"
-    assert key_value_result["value_type"] == "UserSchema"
-    assert key_value_result["value_name"] == "John Doe"
-    assert key_value_result["value_age"] == 30
-
-
-def test_kafka_consumer_with_different_serializers_for_key_and_value(lambda_context):
-    """Test using different serializer types for key and value."""
-
-    # Create JSON data for key and value
-    key_data = {"id": "user-456"}
-    value_data = {"name": "Alice Wonder", "age": 35}
-
-    # Encode the data
-    key_encoded = base64.b64encode(json.dumps(key_data).encode("utf-8")).decode("utf-8")
-    value_encoded = base64.b64encode(json.dumps(value_data).encode("utf-8")).decode("utf-8")
-
-    # Create a kafka event
-    event = {
-        "eventSource": "aws:kafka",
-        "records": {
-            "my-topic-1": [
-                {
-                    "topic": "my-topic-1",
-                    "partition": 0,
-                    "offset": 15,
-                    "timestamp": 1545084650987,
-                    "timestampType": "CREATE_TIME",
-                    "key": key_encoded,
-                    "value": value_encoded,
-                    "headers": [],
-                },
-            ],
-        },
-    }
-
-    # Define models
-    class KeyModel(BaseModel):
-        id: str
-
-    @dataclass
-    class ValueDataClass:
-        name: str
-        age: int
-
-    # Create dict to capture results
-    results = {}
-
-    # Create schema config with different serializers
-    schema_config = SchemaConfig(
-        value_schema_type="JSON",
-        value_output_serializer=ValueDataClass,
-        key_schema_type="JSON",
-        key_output_serializer=KeyModel,
-    )
-
-    @kafka_consumer(schema_config=schema_config)
-    def handler(event: ConsumerRecords, context):
-        record = next(event.records)
-        results["key_type"] = type(record.key).__name__
-        results["key_id"] = record.key.id
-        results["value_type"] = type(record.value).__name__
-        results["value_name"] = record.value.name
-        results["value_age"] = record.value.age
-        return {"processed": True}
-
-    # Call the handler
-    result = handler(event, lambda_context)
-
-    # Verify the results
-    assert result == {"processed": True}
-    assert results["key_type"] == "KeyModel"
-    assert results["key_id"] == "user-456"
-    assert results["value_type"] == "ValueDataClass"
-    assert results["value_name"] == "Alice Wonder"
-    assert results["value_age"] == 35
-
-
 def test_kafka_consumer_without_schema_config(kafka_event_with_json_data, lambda_context):
     """Test Kafka consumer when no schema config is provided."""
 
@@ -847,44 +233,3 @@ def test_kafka_consumer_without_schema_config(kafka_event_with_json_data, lambda
     # Verify the results
     assert result == {"processed": True}
     assert result_data["value_type"] == "str"  # Raw base64 string
-
-
-def test_kafka_consumer_with_custom_dict_class(kafka_event_with_json_data, lambda_context):
-    """Test Kafka consumer with custom dict-like class with to_dict/from_dict methods."""
-
-    # Create a custom dict-like class
-
-    class CustomDict:
-        def __init__(self, **kwargs):
-            for key, value in kwargs.items():
-                setattr(self, key, value)
-
-        @classmethod
-        def from_dict(cls, data):
-            return cls(**data)
-
-        def to_dict(self):
-            return dict(self.__dict__.items())
-
-    # Create dict to capture results
-    result_data = {}
-
-    schema_config = SchemaConfig(value_schema_type="JSON", value_output_serializer=CustomDict)
-
-    @kafka_consumer(schema_config=schema_config)
-    def handler(event: ConsumerRecords, context):
-        # Capture the results to verify
-        record = next(event.records)
-        result_data["value_type"] = type(record.value).__name__
-        result_data["name"] = record.value.name
-        result_data["age"] = record.value.age
-        return {"processed": True}
-
-    # Call the handler
-    result = handler(kafka_event_with_json_data, lambda_context)
-
-    # Verify the results
-    assert result == {"processed": True}
-    assert result_data["value_type"] == "CustomDict"
-    assert result_data["name"] == "John Doe"
-    assert result_data["age"] == 30
